backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.0",
    "fastapi[standard]>=0.115.8",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.0",
    "python-dotenv>=1.0.0", # For environment variables
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.37.0",
]
